+++ examples/protobuf/compiler.nix
@@ -2,4 +2,5 @@
 base.callFile ../../languages/rust/protobuf/compiler {
   mkClient = base.languages.rust.mkClient;
   protobuf = pkgs.protobuf;
+  tonicBuildVersion = "=${base.versions.tonicBuild}";
 }