+++ languages/rust/protobuf/compiler/changeTonicBuildVersion.sh
@@ -1,0 +1,5 @@
+changeTonicBuildVersion() {
+    sed -i 's/tonic-build.*$/tonic-build = "@tonicBuildVersion@"/g' Cargo.toml
+}
+
+preConfigureHooks+=(changeTonicBuildVersion)